Taking in the date created from the file. Closes #3.

+++ MediaOrganiser/Medium/IMedium.cs
@@ -17,7 +17,7 @@ namespace MediaOrganiser
 
         DateTime DateTaken { get; }
 
-        long Length { get ; }
+        double Length { get ; }
 
         bool IsProcessed { get; }
     }
